requirements.txt
# Updated requirements.txt for Python 3.13 compatibility
# Core dependencies
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.10.0
billiard==4.2.1
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.3.2
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
comtypes==1.4.10
cryptography==45.0.4
cssselect==1.3.0
dataclasses-json==0.6.7
DataRecorder==3.6.2
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.12.0
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
docopt==0.6.2
DownloadKit==2.0.7
DrissionPage==4.1.0.18
easyocr==1.7.2
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
facebook-sdk==3.1.0
# Updated numpy for Python 3.13 compatibility
numpy>=1.26.0
Faker==37.4.0
fastapi==0.104.1
filelock==3.17.0
firebase-admin==6.2.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.6.0
fsspec==2025.2.0
future==1.0.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-api-python-client==2.108.0
google-auth==2.29.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-generativeai==0.3.2
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.2.3
groq==0.28.0
grpcio==1.71.0
grpcio-status==1.63.0rc1
gspread==6.2.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
ipython==8.24.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.1
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.64
langchain-groq==0.3.2
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
langsmith==0.3.45
lazy_loader==0.4
limits==5.4.0
linkedin-api==2.0.0
lxml==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.4.3
mypy-extensions==1.0.0
mysqlclient==2.2.7
networkx==3.0
ngrok==1.4.0
ninja==1.11.1.4
npm==0.1.1
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.4.6
openai==1.65.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
optional-django==0.1.0
ordered-set==4.1.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.0
# Updated pandas for Python 3.13 compatibility
pandas>=2.1.0
parso==0.8.4
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
pefile==2023.2.7
phonenumbers==9.0.5
pillow==10.4.0
pipwin==0.5.2
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.0
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pyclipper==1.3.0.post6
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pyjsparser==2.7.1
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.0
pytest-mock==3.11.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
python-whois==0.7.3
pyttsx3==2.98
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.3.0
regex==2024.11.6
reportlab==4.0.4
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
safetensors==0.5.3
schedule==1.2.0
scikit-image==0.25.2
# Updated scikit-learn for Python 3.13 compatibility
scikit-learn>=1.4.0
# Updated scipy for Python 3.13 compatibility
scipy>=1.12.0
selenium==4.34.2
sendgrid==6.10.0
sentence-transformers==4.1.0
setuptools>=80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starkbank-ecdsa==2.2.0
starlette==0.27.0
stripe==7.13.0
sympy==1.14.0
tenacity==9.0.0
tesseract==0.1.3
threadpoolctl==3.6.0
tiktoken==0.8.0
tldextract==5.3.0
tokenizers==0.21.0
tomli==2.2.1
# Updated torch for Python 3.13 compatibility
torch>=2.2.0
torchvision>=0.17.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.5.0
user-agents==2.2.0
uv==0.7.14
uvicorn==0.24.0.post1
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
wheel>=0.45.1
whitenoise==6.6.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0 